# doc/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: Gestionnaire de projets
  version: 1.0.0
  title: CDP
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
host: virtserver.swaggerhub.com
basePath: /edupland/CDP/1.0.0
...
schemes:
- https
paths:
...
  /issue:
    get:
      tags:
      - Issue
      summary: Récupération des issues
      description: On récupère les issues d'un projet
      parameters:
      - name: IDProjet
        in: query
        description: Id du projet dont on veut les issues
        required: true
        type: string
      responses:
        "200":
          description: La liste des issues du projet dont l'id est IDProjet
        "500":
          description: |-
            Connexion à la base de données impossible
            Échec de la récupération des issues
    post:
      tags:
      - Issue
      summary: Création d'une issue
      description: Une issue est créée
      parameters:
      - name: IDProjet
        in: query
        description: Id du projet
        required: true
        type: string
      - name: Role
        in: query
        description: Role de l'issue
        required: true
        type: string
      - name: Action
        in: query
        description: Action de l'issue
        required: true
        type: string
      - name: Raison
        in: query
        description: Raison de l'issue
        required: true
        type: string
      responses:
        "201":
          description: Création de l'issue réussie
          schema:
            $ref: '#/definitions/Issue'
        "500":
          description: |-
            Connexion à la base de données impossible
            Échec de la création de l'issue
...
  /test:
    get:
      tags:
      - Test
      summary: Récupération des tests
      description: On récupère la liste des tests d'un projet
      parameters:
      - name: IDProjet
        in: query
        description: Id du projet dont on veut les tests
        required: true
        type: string
      responses:
        "200":
          description: La liste des tests du projet dont l'id est IDProjet
        "500":
          description: |-
            Connexion à la base de données impossible
            Échec de la récupération des tests
    post:
      tags:
      - Test
      summary: Création d'un test
      description: Un test est créé
      parameters:
      - name: IDProjet
        in: query
        description: Id du projet
        required: true
        type: string
      - name: Given
        in: query
        description: Initialisation du test
        required: true
        type: string
      - name: When
        in: query
        description: Élément qui déclenche un évènement
        required: true
        type: string
      - name: Then
        in: query
        description: Effet de l'évènement
        required: true
        type: string
      responses:
        "200":
          description: Création d'un test réussie
          schema:
            $ref: '#/definitions/Test'
        "500":
          description: |-
            Connexion à la base de données impossible
            Échec de la création d'un test
definitions:
...
  Issue:
    type: object
    required:
    - Action
    - Projet
    - Raison
    - Role
    properties:
      Projet:
        $ref: '#/definitions/Issue_Projet'
      Role:
        type: string
        example: développeur
      Action:
        type: string
        example: avoir accès à la base de données
      Raison:
        type: string
        example: pouvoir modifier les informations d'un utilisateur
...
  Test:
    type: object
    required:
    - Projet
    - testGiven
    - testThen
    - testWhen
    properties:
      Projet:
        $ref: '#/definitions/Issue_Projet'
      testGiven:
        type: string
        example: Je crée un projet
      testWhen:
        type: string
        example: Je ne suis pas connecté à l'outil
      testThen:
        type: string
        example: Je vois une alerte "Pas connecté"
  Project_Proprietaire:
    type: object
    required:
    - Login
    - Mail
    - Nom
    - Password
    - Prenom
    properties:
      Nom:
        type: string
        example: Dupont
      Prenom:
        type: string
        example: Michel
      Mail:
        type: string
        example: michel.dupont@mail.fr
      Societe:
        type: string
      Login:
        type: string
        example: mdupont
      Password:
        type: string
        example: mdupont
  Project_Developpeurs:
    type: object
    required:
    - Login
    - Mail
    - Nom
    - Password
    - Prenom
    properties:
      Nom:
        type: string
        example: Garcia
      Prenom:
        type: string
        example: Gabriel
      Mail:
        type: string
        example: gabriel.garcia@mail.fr
      Societe:
        type: string
        example: Capgemini
      Login:
        type: string
        example: ggarcia
      Password:
        type: string
        example: ggarcia
  Project_Clients:
    type: object
    required:
    - Login
    - Mail
    - Nom
    - Password
    - Prenom
    properties:
      Nom:
        type: string
        example: Martin
      Prenom:
        type: string
        example: Bernard
      Mail:
        type: string
        example: bernard.martin@mail.fr
      Societe:
        type: string
        example: Dassault
      Login:
        type: string
        example: bmartin
      Password:
        type: string
        example: bmartin
  Issue_Projet:
    type: object
    required:
    - Clients
    - Description
    - Developpeurs
    - Proprietaire
    - Titre
    properties:
      Titre:
        type: string
        example: Titre du projet
      Description:
        type: string
        example: Voici la description du projet
      Proprietaire:
        $ref: '#/definitions/Project_Proprietaire'
      Developpeurs:
        $ref: '#/definitions/Project_Developpeurs'
      Clients:
        $ref: '#/definitions/Project_Clients'